+++ src/nludb/types/base.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+from inspect import currentframe, getargvalues
+from typing import Type
+
+@dataclass
+class NludbRequest():
+  pass
+
+@dataclass
+class NludbResponse():
+  pass